requirements.txt
yfinance==0.1.59
Django==3.2.2
djangorestframework==3.12.4
